+++ servicio-transporte/src/main/java/com/backend/tpi_backend/servicio_transporte/repositories/RutaRepository.java
@@ -5,5 +5,6 @@ import java.util.List;
 import com.backend.tpi_backend.servicio_transporte.model.Ruta;
 
 public interface RutaRepository extends BaseRepository<Ruta, Integer> {
-    List<Ruta> findBySolicitudId(Long solicitudId);
+    // Filtra las rutas que pertenecen a una solicitud específica
+    List<Ruta> findBySolicitudId(Integer solicitudId);
 }
